se agrega camaraScreen para manejar permisos al ingresar km inicial
