+++ android/app/build.gradle
@@ -151,6 +151,12 @@ android {
             keyAlias 'androiddebugkey'
             keyPassword 'android'
         }
+     release {
+          storeFile file('AppTrackKey.keystore')
+         storePassword 'G3nerico01'
+         keyAlias 'AppTrackAlias'
+         keyPassword 'G3nerico01'
+    }
 
     }
     buildTypes {
@@ -160,7 +166,9 @@ android {
         release {
             // Caution! In production, you need to generate your own keystore file.
             // see https://reactnative.dev/docs/signed-apk-android.
-            signingConfig signingConfigs.debug
+            // signingConfig signingConfigs.debug
+             signingConfig signingConfigs.release
+
             minifyEnabled enableProguardInReleaseBuilds
             proguardFiles getDefaultProguardFile("proguard-android.txt"), "proguard-rules.pro"
         }
